4º Course/First Semester/GESTION DE PROYECTOS INFORMATICOS I/g-8/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.10
pytz==2024.2
sqlparse==0.5.1
whitenoise==6.7.0
